I have just realized the perfect data structure for the information acquired through this project. The data structure is a list. The list can be represented in a row or a column in excel. The format of the list is  `<benefiary>; <rating>; <explanation>; <attribute 1>;...; <attribute n>[;]`
Additional attributes can be added even after <rating> and <explanation> have been added by always defining the second and third items as <rating> and <explanation>. This is the unix way!!! My datastructure is representable as a textual object now because of its single-dimensional nature. The list is independent of implementation: it can be stored as an excel row or column, a database relation or other structure, a csv, etc.

Implementation as a column in Excel:
1. name `<beneficiary>` in a cell
1. placeholder of `"rating"` goes in cell below
1. placeholder of `"explanation"` goes in cell below
1. each cell below contains an attribute